Renaming template.
