+++ app/build.gradle.kts
@@ -12,7 +12,7 @@ android {
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
-        applicationId = "template.app.id"
+        applicationId = "com.adammcneilly.spacenerd"
         minSdk = libs.versions.minSdk.get().toInt()
         targetSdk = libs.versions.compileSdk.get().toInt()
         versionCode = 1
@@ -50,7 +50,7 @@ android {
         }
     }
 
-    namespace = "template"
+    namespace = "com.adammcneilly.spacenerd"
 }
 
 dependencies {
