+++ app/src/main/java/com/adammcneilly/spacenerd/MainActivity.kt
@@ -1,4 +1,4 @@
-package template
+package com.adammcneilly.spacenerd
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,8 +8,8 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import com.adammcneilly.spacenerd.theme.SpaceTheme
 import dagger.hilt.android.AndroidEntryPoint
-import template.theme.TemplateTheme
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -19,7 +19,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             enableEdgeToEdge()
 
-            TemplateTheme {
+            SpaceTheme {
                 Surface(
                     color = MaterialTheme.colorScheme.background,
                 ) {
