+++ app/src/main/java/com/adammcneilly/spacenerd/SpaceNerdApp.kt
@@ -1,7 +1,7 @@
-package template
+package com.adammcneilly.spacenerd
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class TemplateApp : Application()
+class SpaceNerdApp : Application()
