+++ app/src/main/java/com/adammcneilly/spacenerd/theme/Shape.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.spacenerd.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
