+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">template</string>
+    <string name="app_name">Space Nerd</string>
 </resources>
